requirements.txt
storops
hurry.filesize
